+++ functions/image-lazyload-native.php
@@ -5,7 +5,7 @@
  * @Author:		Elias Kautto
  * @Date:   		2022-01-28 12:33:30
  * @Last Modified by:   Elias Kautto
- * @Last Modified time: 2022-01-28 12:51:43
+ * @Last Modified time: 2022-02-01 10:24:12
  *
  * @package air-helper
  */
@@ -14,13 +14,14 @@
  * Echo image in lazyloading tag for native-lazyload.
  *
  * @param  integer $image_id Image attachment id to lazyload.
- * @param  string  $fallback Url for fallback image. Defaults to theme settings default featured image.
- * @param  array   $sizes    Custom sizes for lazyloading. Optional.
+ * @param  string  $args['fallback'] Url for fallback image. Defaults to theme settings default featured image.
+ * @param  array   $args['sizes']    Custom sizes for lazyloading. Optional.
+ * @param  string  $args['class']    Class name to give for img tag. Optional.
  * @since  2.3.1
  */
 if ( ! function_exists( 'native_lazyload_tag' ) ) {
-  function native_lazyload_tag( $image_id = 0, $fallback = false, $sizes = [] ) {
-    echo get_native_lazyload_tag( $image_id, $fallback, $sizes ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+  function native_lazyload_tag( $image_id = 0, $args = [] ) {
+    echo get_native_lazyload_tag( $image_id, $args ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
   } // end native_lazyload_tag
 } // end if
 
@@ -28,19 +29,26 @@ if ( ! function_exists( 'native_lazyload_tag' ) ) {
  * Get image lazyloading tag for native-lazyload.
  *
  * @param  integer $image_id Image attachment id to lazyload.
- * @param  string  $fallback Url for fallback image. Defaults to theme settings default featured image.
- * @param  array   $sizes    Custom sizes for lazyloading. Optional.
+ * @param  string  $args['fallback'] Url for fallback image. Defaults to theme settings default featured image.
+ * @param  array   $args['sizes']    Custom sizes for lazyloading. Optional.
+ * @param  string  $args['class']    Class name to give for img tag. Optional.
  * @return string            String containing lazyloading tag.
  * @since  2.3.1
  */
 if ( ! function_exists( 'get_native_lazyload_tag' ) ) {
-  function get_native_lazyload_tag( $image_id = 0, $fallback = false, $sizes = [] ) {
-    // Get image
-    $image_urls = air_helper_get_image_lazyload_sizes( $image_id, $sizes );
+  function get_native_lazyload_tag( $image_id = 0, $args = [] ) {
+    $args = wp_parse_args( $args, [
+      'fallback' => false,
+      'sizes' => [],
+      'class' => null,
+    ] );
+
+  // Get image
+    $image_urls = air_helper_get_image_lazyload_sizes( $image_id, $args['sizes'] );
 
     // Check if we have image
     if ( ! $image_urls || ! is_array( $image_urls ) ) {
-      return get_native_lazyload_tag_fallback( $fallback );
+      return get_native_lazyload_tag_fallback( $args['fallback'] );
     }
 
     // Possibility to add optional styles for the image
@@ -56,11 +64,15 @@ if ( ! function_exists( 'get_native_lazyload_tag' ) ) {
 
     // Get the img tag
     ob_start(); ?>
-    <img class="lazy"
-      loading="lazy"
+    <img loading="lazy"
       alt="<?php echo esc_attr( $alt ); ?>"
-      style="<?php echo esc_attr( $styles ); ?>"
       src="<?php echo esc_url( $image_urls['big'] ); ?>"
+      <?php if ( ! empty( $styles ) ) : ?>
+        style="<?php echo esc_attr( $styles ); ?>"
+      <?php endif; ?>
+      <?php if ( ! empty( $args['class'] ) ) : ?>
+        class="<?php echo esc_attr( $args['class'] ); ?>"
+      <?php endif; ?>
     />
     <?php
 
@@ -71,34 +83,36 @@ if ( ! function_exists( 'get_native_lazyload_tag' ) ) {
 /**
  * Fallback for lazyload tag.
  *
- * @param string $fallback Url for fallback image. Defaults to theme settings default featured image.
+ * @param string $args['fallback'] Url for fallback image. Defaults to theme settings default featured image.
+ * @param  string  $args['class']    Class name to give for img tag. Optional.
  * @return string            String containing lazyloading tag.
  * @since 2.4.0
  */
 if ( ! function_exists( 'get_native_lazyload_tag_fallback' ) ) {
-  function get_native_lazyload_tag_fallback( $fallback = false ) {
+  function get_native_lazyload_tag_fallback( $args ) {
     // Default to theme default featured image if no fallback specified
-    if ( empty( $fallback ) ) {
+    if ( empty( $args['fallback'] ) ) {
       if ( apply_filters( 'air_helper_image_lazyload_fallback_from_theme_settings', true ) && defined( 'THEME_SETTINGS' ) ) {
-        $fallback = THEME_SETTINGS['default_featured_image'];
+        $args['fallback'] = THEME_SETTINGS['default_featured_image'];
       }
     }
 
     // No fallback, bail
-    if ( empty( $fallback ) ) {
+    if ( empty( $args['fallback'] ) ) {
       return;
     }
 
     // Get the img tag
     ob_start(); ?>
-    <img class="lazy"
-      loading="lazy"
+    <img loading="lazy"
       alt=""
-      src="<?php echo esc_url( $fallback ); ?>"
+      src="<?php echo esc_url( $args['fallback'] ); ?>"
+      <?php if ( ! empty( $args['class'] ) ) : ?>
+        class="<?php echo esc_attr( $args['class'] ); ?>"
+      <?php endif; ?>
     />
     <?php
 
     return ob_get_clean();
   } // end get_native_lazyload_tag_fallback
 } // end if
-